+++ ImTK/UI/Element/Basic/RectField.cs
@@ -22,7 +22,7 @@ namespace ImTK.UI
             classList.Add("rect-field");
         }
 
-        protected override void OnRenderSelf()
+        public override void OnRender()
         {
             System.Numerics.Vector4 currentValues = new System.Numerics.Vector4(value.x, value.y, value.width, value.height);
             string format = $"%.{precision}f";